# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Homewarp",
    "description": "The Homewarp panel's API.",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/v1/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "401": {
            "description": "Wrong username or password.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/setup": {
      "post": {
        "summary": "Creates the first account and signs it in. Works once.",
        "operationId": "setup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The account exists and is signed in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "403": {
            "description": "The setup code is wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            }
          },
          "409": {
            "description": "There is an account already.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            }
          },
          "422": {
            "description": "The username or password will not do.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
      "ProblemBody": {
        "type": "object",
        "description": "The body of every error response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "A sentence fit to show to the person using the panel."
          }
        }
      },
...
      "Session": {
        "type": "object",
        "description": "Everything the web interface needs to decide what to show first.",
        "required": [
          "setup_required",
          "version"
        ],
        "properties": {
          "setup_required": {
            "type": "boolean",
            "description": "True until the first account has been created."
          },
          "sftp_port": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "The port SFTP is reached on at this machine, where it is served.",
            "minimum": 0
          },
          "user": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/User",
                "description": "Who is signed in, if anyone."
              },
              {
                "type": "null"
              }
            ]
          },
          "version": {
            "type": "string"
          }
        }
      },
...
      "SetupRequest": {
        "type": "object",
        "required": [
          "code",
          "username",
          "password"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "The code Homewarp wrote to its log when it started."
          },
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
      "User": {
        "type": "object",
        "description": "An account.",
        "required": [
          "id",
          "username",
          "owner"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "owner": {
            "type": "boolean",
            "description": "The account made at setup. It may do everything; any other account\nonly what it has been let do, with the servers it has been let into."
          },
          "username": {
            "type": "string"
          }
        }
      }
    }
  }
}